+++ src/modules/utils/CMakeLists.txt
@@ -16,11 +16,11 @@ forte_add_module(UTILS OFF "FORTE UTILITY FBs")
 # FORTE UTILITY FBs
 #############################################################################
 
-forte_add_sourcefile_hcpp(E_STOPWATCH GetInstancePath_fbt GetInstancePathAndName_fbt)
-forte_add_sourcefile_hcpp(OUT_ANY_CONSOLE GEN_F_MUX GEN_CSV_WRITER)
-forte_add_sourcefile_hcpp(GEN_ARRAY2VALUES GEN_VALUES2ARRAY GEN_ARRAY2ARRAY GET_AT_INDEX SET_AT_INDEX)
-forte_add_sourcefile_hcpp(FB_RANDOM GET_STRUCT_VALUE SET_STRUCT_VALUE)
+forte_add_sourcefile_hcpp(E_STOPWATCH GetInstancePath_fbt_fbt GetInstancePathAndName_fbt)
+forte_add_sourcefile_hcpp(OUT_ANY_CONSOLE_fbt GEN_F_MUX_fbt GEN_CSV_WRITER_fbt)
+forte_add_sourcefile_hcpp(GEN_ARRAY2VALUES_fbt GEN_VALUES2ARRAY_fbt GEN_ARRAY2ARRAY_fbt GET_AT_INDEX_fbt SET_AT_INDEX_fbt)
+forte_add_sourcefile_hcpp(FB_RANDOM GET_STRUCT_VALUE_fbt SET_STRUCT_VALUE_fbt)
 
-forte_add_sourcefile_hcpp(STEST_END)
-forte_add_sourcefile_hcpp(TEST_CONDITION)
+forte_add_sourcefile_hcpp(STEST_END_fbt)
+forte_add_sourcefile_hcpp(TEST_CONDITION_fbt)
 forte_add_subdirectory(Selection)
