+++ src/modules/utils/SET_AT_INDEX_fbt.cpp
@@ -14,9 +14,9 @@
  *     - refactor for ANY variant
  *******************************************************************************/
 
-#include "SET_AT_INDEX.h"
+#include "SET_AT_INDEX_fbt.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "SET_AT_INDEX_gen.cpp"
+#include "SET_AT_INDEX_fbt_gen.cpp"
 #endif
 
 #include "criticalregion.h"
